Added notes to readme, moved Digestible to a digest package
touchups
