+++ project/WebsitePlugin.scala
@@ -16,6 +16,6 @@ object WebsitePlugin extends AutoPlugin {
       micrositeDocumentationUrl := "docs",
       micrositeGithubToken := Option(System.getenv().get("GITHUB_TOKEN")),
       micrositePushSiteWith := GitHub4s,
-      micrositeHighlightLanguages ++= Seq("java", "scala", "kotlin")
+      micrositeHighlightLanguages ++= Seq("java", "scala", "kotlin", "groovy")
     )
 }
